refactor: Move each form to its own component and adjust error handling
